feat: wip: new home page
