+++ lib/app/home_old/_widgets/location_not_set_card.dart
@@ -1,0 +1,56 @@
+/// Card shown when no home location has been configured.
+library;
+
+import 'package:dpip/core/i18n.dart';
+import 'package:dpip/router.dart';
+import 'package:dpip/utils/extensions/build_context.dart';
+import 'package:flutter/material.dart';
+import 'package:material_symbols_icons/material_symbols_icons.dart';
+
+/// Prompts the user to set a home location when none is configured.
+///
+/// Displays an icon, explanatory text, and a button that navigates to the
+/// location settings page.
+class LocationNotSetCard extends StatelessWidget {
+  /// Creates a [LocationNotSetCard].
+  const LocationNotSetCard({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        color: context.colors.secondaryContainer,
+        border: Border.all(color: context.colors.secondary, width: 2),
+        borderRadius: .circular(16),
+      ),
+      padding: const .only(left: 12, top: 4, right: 4, bottom: 4),
+      clipBehavior: Clip.antiAlias,
+      child: Row(
+        mainAxisSize: .min,
+        mainAxisAlignment: .spaceBetween,
+        children: [
+          Row(
+            spacing: 8,
+            children: [
+              Icon(
+                Symbols.not_listed_location_rounded,
+                color: context.colors.onSecondaryContainer,
+                weight: 500,
+              ),
+              Text(
+                '尚未設定所在地'.i18n,
+                style: context.texts.bodyMedium!.copyWith(
+                  color: context.colors.onSecondaryContainer,
+                ),
+              ),
+            ],
+          ),
+          TextButton(
+            child: Text('設定'.i18n),
+            onPressed: () => SettingsLocationRoute().push(context),
+          ),
+        ],
+      ),
+    );
+  }
+}
