+++ lib/widgets/typography.dart
@@ -31,9 +31,18 @@ class DisplayText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     TextStyle? style,
     this.align,
-  }) : style = TextStyle(color: color, fontWeight: weight).merge(style),
+  }) : style = TextStyle(
+         color: color,
+         fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
+       ).merge(style),
        _textStyleGetter = ((context) => context.texts.displaySmall);
 
   /// Creates a medium display text widget.
@@ -45,9 +54,18 @@ class DisplayText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     TextStyle? style,
     this.align,
-  }) : style = TextStyle(color: color, fontWeight: weight).merge(style),
+  }) : style = TextStyle(
+         color: color,
+         fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
+       ).merge(style),
        _textStyleGetter = ((context) => context.texts.displayMedium);
 
   /// Creates a large display text widget.
@@ -59,9 +77,18 @@ class DisplayText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     TextStyle? style,
     this.align,
-  }) : style = TextStyle(color: color, fontWeight: weight).merge(style),
+  }) : style = TextStyle(
+         color: color,
+         fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
+       ).merge(style),
        _textStyleGetter = ((context) => context.texts.displayLarge);
 
   @override
@@ -103,9 +130,18 @@ class HeadLineText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     TextStyle? style,
     this.align,
-  }) : style = TextStyle(color: color, fontWeight: weight).merge(style),
+  }) : style = TextStyle(
+         color: color,
+         fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
+       ).merge(style),
        _textStyleGetter = ((context) => context.texts.headlineSmall);
 
   /// Creates a medium headline text widget.
@@ -117,9 +153,18 @@ class HeadLineText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     TextStyle? style,
     this.align,
-  }) : style = TextStyle(color: color, fontWeight: weight).merge(style),
+  }) : style = TextStyle(
+         color: color,
+         fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
+       ).merge(style),
        _textStyleGetter = ((context) => context.texts.headlineMedium);
 
   /// Creates a large headline text widget.
@@ -131,9 +176,18 @@ class HeadLineText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     TextStyle? style,
     this.align,
-  }) : style = TextStyle(color: color, fontWeight: weight).merge(style),
+  }) : style = TextStyle(
+         color: color,
+         fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
+       ).merge(style),
        _textStyleGetter = ((context) => context.texts.headlineLarge);
 
   @override
@@ -176,12 +230,18 @@ class TitleText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.titleSmall);
@@ -195,12 +255,18 @@ class TitleText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.titleMedium);
@@ -214,12 +280,18 @@ class TitleText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.titleLarge);
@@ -264,12 +336,18 @@ class BodyText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.bodySmall);
@@ -283,12 +361,18 @@ class BodyText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.bodyMedium);
@@ -302,12 +386,18 @@ class BodyText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.bodyLarge);
@@ -352,12 +442,18 @@ class LabelText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.labelSmall);
@@ -371,12 +467,18 @@ class LabelText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.labelMedium);
@@ -390,12 +492,18 @@ class LabelText extends StatelessWidget {
     super.key,
     Color? color,
     FontWeight? weight,
+    String? fontFamily,
+    double? fontSize,
+    List<FontVariation>? fontVariations,
     double? leading,
     TextStyle? style,
     this.align,
   }) : style = TextStyle(
          color: color,
          fontWeight: weight,
+         fontFamily: fontFamily,
+         fontSize: fontSize,
+         fontVariations: fontVariations,
          height: leading,
        ).merge(style),
        _textStyleGetter = ((context) => context.texts.labelLarge);
@@ -408,4 +516,8 @@ class LabelText extends StatelessWidget {
       textAlign: align,
     );
   }
+}
+
+extension FontVariationExtension on FontVariation {
+  static roundness(double round) => FontVariation('ROND', round);
 }
